feat: fetch article detail
