--- src/app/api/articles/detail/[slug]/route.ts
@@ -1,23 +1,0 @@
-import { NextResponse } from 'next/server';
-
-export async function GET(request: Request, { params }: { params: { slug: string } }) {
-	const slug = (await params).slug;
-
-	try {
-		console.log('params:', slug);
-		const res = await fetch(`http://localhost:1000/articles/detail/${slug}`, {
-			headers: {
-				'Content-Type': 'application/json',
-				Accept: 'application/json',
-			},
-		});
-		if (!res.ok) {
-			throw new Error('Failed to fetch data');
-		}
-		const data = await res.json();
-		return NextResponse.json(data);
-	} catch (err) {
-		console.error(err);
-		return NextResponse.error();
-	}
-}
