+++ src/app/articles/detail/[slug]/_components/PrevNextArticle.tsx
