+++ src/app/articles/detail/[slug]/_components/content/SubscribeInduce.tsx
