--- src/components/article/content/ArticleContent.tsx
@@ -1,96 +1,0 @@
-import styled from "styled-components";
-import DefaultImg from "@/components/common/DefaultImg";
-import SubscribeInduce from "@/components/article/content/SubscribeInduce";
-
-interface NewsletterContentProps {
-  flex?: number;
-  children?: React.ReactNode;
-  className?: string;
-}
-
-function ArticleContent({flex, className}: NewsletterContentProps) {
-  return (
-    <ContainerStyled flex={flex} className={className}>
-      <NewsletterContentStyled>
-        <DefaultImg />
-        <h1>1 hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</h1>
-        <p>hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</p>
-        <h2>2 hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</h2>
-        <p>hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</p>
-        <h3>3 hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</h3>
-        <p>hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</p>
-        <h4>4 hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</h4>
-        <p>hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</p>
-        <h5>5 hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</h5>
-        <p>hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</p>
-        <h6>6 hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</h6>
-        <p>hellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohellohello</p>
-        <figure>
-
-        </figure>
-        <a href="#">https://notefolio.net/service/content/post/86540</a>
-        <strong>hellllllllop</strong>
-        <em>hello world</em>
-        <div className="content-bottom-margin" />
-      </NewsletterContentStyled>
-      <SubscribeInduce />
-    </ContainerStyled>
-  );
-}
-
-const ContainerStyled = styled.div<NewsletterContentProps>`
-    position: relative;
-    display: flex;
-    flex-direction: column;
-`
-
-const NewsletterContentStyled = styled.article`
-    position: relative;
-    gap: 0.5rem;
-
-    .content-bottom-margin {
-        padding-bottom: 2rem;
-    }
-
-    * {
-        margin-bottom: 0.5rem;
-        word-break: break-word;
-        line-height: 1.56;
-    }
-
-    h1, h2, h3, h4, h5, h6 {
-        margin-top: 2rem;
-    }
-
-    p {
-        color: ${({theme}) => theme.color.mediumGrey};
-    }
-
-    img {
-        border-radius: ${({theme}) => theme.borderRadius.soft};
-    }
-
-    figure {
-        margin: 0;
-        width: 100%;
-
-        img {
-            margin-bottom: 0.25rem;
-            width: 100%;
-        }
-
-        figcaption {
-            font-size: ${({theme}) => theme.fontSize.extraSmall};
-            color: ${({theme}) => theme.color.lightGrey};
-        }
-    }
-
-    a {
-        color: ${({theme}) => theme.color.primary};
-        text-decoration: underline ${({theme}) => theme.color.primary};
-        text-decoration-thickness: 2px;
-    }
-`;
-
-
-export default ArticleContent;
