+++ src/utils/stripCodeFence.ts
@@ -1,0 +1,10 @@
+export function stripCodeFence(content: string, language: string = 'html'): string {
+  const codeFenceStart = `\`\`\`${language}\n`;
+  const codeFenceEnd = '```';
+
+  if (content.startsWith(codeFenceStart) && content.endsWith(codeFenceEnd)) {
+    return content.slice(codeFenceStart.length, -codeFenceEnd.length).trim();
+  }
+
+  return content;
+}
